# api/openapi/kapsora-v1.yaml
openapi: 3.1.0
info:
  title: KAPSORA Platform API
  version: 1.0.0-alpha.1
  summary: Kurumsal hak, fayda ve hizmet orkestrasyonu için başlangıç API sözleşmesi
  description: |
    Bu sözleşme ilk sprint ve çekirdek MVP için contract-first geliştirme başlangıç noktasıdır.
    Kullanıcılar KAPSORA'nın kendi hesaplarıyla giriş yapar (ADR-022); tarayıcı yalnız opak
    oturum çerezi taşır. Makineden makineye entegrasyonlar OAuth 2.0 client credentials ve
    mTLS ile çalışır. Tüm tenant kapsamlı çağrılar X-Tenant-ID
    başlığı taşır. Durum değiştiren POST komutlarında Idempotency-Key zorunludur; sorgu
    niteliğindeki POST'larda (eligibility check, pricing quote) opsiyoneldir.
    İstemci X-Request-ID başlığında UUID gönderebilir; sunucu her yanıtta X-Request-ID döner.
    Çerez oturumuyla yapılan durum değiştiren çağrılar X-CSRF-Token başlığı taşır.
  contact:
    name: KAPSORA Product Engineering
servers:
  - url: https://api.staging.kapsora.example
    description: Staging
  - url: https://api.kapsora.example
    description: Production
security:
  - sessionCookie: []
  - oauth2: [kapsora.api]
...
paths:
...
  /api/v1/contracts:
    get:
      tags: [Contracts]
      operationId: listContracts
      description: |
        Contracts of the tenant, newest first, with keyset paging. A contract is the
        agreement between one payer organization and one provider; what was actually
        agreed lives in its versions. q searches the code and the name, the other filters
        narrow to one provider, one payer, one service domain or one status.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: q
          in: query
          description: Contract code or name search.
          schema:
            type: string
            maxLength: 120
        - name: providerProfileId
          in: query
          description: Only contracts with this provider.
          schema:
            type: string
            format: uuid
        - name: payerOrganizationId
          in: query
          description: Only contracts whose payer is this organization relationship.
          schema:
            type: string
            format: uuid
        - name: domainCode
          in: query
          description: Only contracts in this service domain.
          schema:
            $ref: '#/components/schemas/ServiceDomain'
        - name: status
          in: query
          description: Only contracts in this status.
          schema:
            $ref: '#/components/schemas/ContractStatus'
      responses:
        '200':
          description: Contract page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractPage'
        '400':
          description: Cursor invalid
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Contracts]
      operationId: createContract
      description: |
        Opens a contract between a payer organization relationship and a provider profile
        of this tenant. The contract carries no prices of its own: it starts in DRAFT and
        holds versions, and only a published version is ever read by a quote, an
        authorization or a claim. The code is unique inside the tenant, so a second one
        answers 409 CONTRACT_CODE_TAKEN.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: Contract created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/contracts/{contractId}:
    get:
      tags: [Contracts]
      operationId: getContract
      description: |
        One contract of the tenant with its parties and its optimistic concurrency tag.
        The versions are read separately, because a long-running contract accumulates more
        of them than a detail view wants to carry.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: Contract
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Contracts]
      operationId: patchContract
      description: |
        Merge-patch of the name, the sponsor organization and the contract status. The
        code, the payer, the provider and the service domain are absent: they are what the
        contract is, and every published version was agreed under them, so a body carrying
        one answers 422 with field code IMMUTABLE. Suspending or closing a contract does
        not change its versions; it only stops the price selection considering them.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateContractRequest'
      responses:
        '200':
          description: Contract updated
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '415':
          description: Body is not application/merge-patch+json
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contracts/{contractId}/versions:
    get:
      tags: [Contracts]
      operationId: listContractVersions
      description: |
        The version summaries of one contract, highest version number first. The price
        sheet of a version is read through the version itself; this list is what a screen
        shows to pick between "what we agreed then" and "what we agreed now".
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: Contract versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersionList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Contracts]
      operationId: createContractVersion
      description: |
        Opens the next DRAFT version of a contract. copyFromVersionId copies the price
        lists, price items, package definitions, quotas and payment term of another
        version of the same contract, which is how a yearly price revision starts from
        last year's sheet instead of from an empty page. The new version stays editable
        until it is published.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractVersionRequest'
      responses:
        '201':
          description: Contract version created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/contract-versions/{contractVersionId}:
    get:
      tags: [Contracts]
      operationId: getContractVersion
      description: |
        One contract version with its price lists, its review metadata and its
        configuration hash. The price items of each list are paged separately, because a
        health tariff runs to thousands of rows. A published version needs only
        `contract.read`; a draft needs `contract.manage`, so a price nobody has agreed to
        yet is not visible to everyone who may read the agreed one.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
      responses:
        '200':
          description: Contract version
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Contracts]
      operationId: patchContractVersion
      description: |
        Merge-patch of the validity period, the currency and the notes of a DRAFT version.
        A version that is not DRAFT answers 409 CONTRACT_VERSION_IMMUTABLE: everything
        downstream was priced from the published sheet, so it cannot move underneath a
        claim that already quoted it.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateContractVersionRequest'
      responses:
        '200':
          description: Contract version updated
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '415':
          description: Body is not application/merge-patch+json
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/submit:
    post:
      tags: [Contracts]
      operationId: submitContractVersion
      description: |
        Moves a DRAFT version to UNDER_REVIEW and records the maker. A version with no
        price list holding at least one price item is refused with 422 on field
        priceLists: an empty sheet cannot be reviewed, and publishing it would make every
        later price lookup answer PRICE_NOT_FOUND. The validity start is required too,
        because a published version without one could never be selected by service date.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewComment'
      responses:
        '200':
          description: Contract version submitted for review
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/publish:
    post:
      tags: [Contracts]
      operationId: publishContractVersion
      description: |
        Freezes a version under review. It needs `contract.publish`, a recent step-up and
        an actor other than the one who submitted it; the same person answers 403
        MAKER_CHECKER_SAME_ACTOR and the refusal is audited. Publishing writes a
        configuration hash over the version's whole price content, so "what was agreed"
        can be proved later. Two published versions of one contract may not cover the same
        date: an overlap answers 409 CONTRACT_VERSION_OVERLAP.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewComment'
      responses:
        '200':
          description: Contract version published
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/retire:
    post:
      tags: [Contracts]
      operationId: retireContractVersion
      description: |
        Closes a published version with a reason. It needs `contract.publish` and a recent
        step-up. The version stays readable for every quote and claim priced from it and
        stops being a candidate for new price lookups; the reason code is required so the
        history says why the tariff was withdrawn.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonCommand'
      responses:
        '200':
          description: Contract version retired
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-kapsora_session
      description: |
        Opaque session id. The server stores only its SHA-256 hash and the browser never
        holds a token of any kind. Local HTTP development uses the name kapsora_session
        because the __Host- prefix requires HTTPS.
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://identity.kapsora.example/realms/kapsora/protocol/openid-connect/token
          scopes:
            kapsora.api: Call KAPSORA APIs
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: Selected tenant UUID. It must be one of the actor's active memberships.
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from the previous response.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
...
    Forbidden:
      description: Authenticated actor lacks permission or scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Request is syntactically valid but violates validation or business preconditions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    ServiceDomain:
      type: string
      description: Closed list of service domains shared by categories and definitions.
      enum: [GENERIC, HEALTH, ACCOMMODATION, ASSISTANCE, EDUCATION, SPORT, TRANSPORT, CARE, OTHER]
...
    ContractStatus:
      type: string
      description: |
        Lifecycle of a contract. Only an ACTIVE contract's published versions are
        considered by the price selection.
      enum: [DRAFT, ACTIVE, SUSPENDED, CLOSED]
...
    ContractPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
        nextCursor:
          type: [string, 'null']
...